==> code/4_FPTree.ipynb <==
# FP Growth Implementation

Applies FP Growth Algorithm on the dataset to generate frequent itemsets.

import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

db_path = os.path.join('..','data','7_Preprocess_Final.csv')
db = pd.read_csv(db_path)
db.head()

### Preprocessing

Drop the binary columns because they are too numerous. We prefer to append the keywords _Rainfall_ and _Yield_ to the data to make it more readable to us.

new_db = db[['Season', 'Crop', 'Soil Type', 'Rainfall_Disc', 'Yield_Disc']]
new_db.head()

x1 = new_db['Rainfall_Disc'].values
x2 = new_db["Yield_Disc"].values
crops = list(set(new_db["Crop"].values))

for i in range(0, len(x1)):
    x1[i] = x1[i]+"_Rainfall"
for i in range(0, len(x2)):
    x2[i] = x2[i]+"_Yield"

new_db.head()

new_db.values

Convert the dataset into a 2D list to ease the data preparation for the FP Tree generation

db = new_db.values.tolist()

## FP Growth Algorithm

Let us now define the classes and the functions required to implement the algorithm

class Node:
    """
    A Node of a tree.

    Default constructor creates a root node. i.e. with everything to be 0
    
    Interior nodes require value and default frequency is 0
    """
    def __init__(self, val=0, freq=0):
        self.val = val
        self.freq = freq
        self.next = []
        self.parent = None
        self.next_occurence = None


class Tree:
    """
    A class to implement an FP-Tree and perform the FP-Growth algorithm to
    mine frequent itemsets.
    
    The class requires preprocessed data, i.e. each transaction in the dataset 
    to be sorted according to the decreasing order of the support counts in 
    the dataset.
    """
    
    def __init__(self, min_sup=1):
        """
        The constructor requires an argument which corresponds to the
        minimum support of the tree
        
        The default constructor creates a tree with minimum support as 1
        """
        self.root = Node()
        self.header = {}
        self.support_count = {}
        self.min_support = min_sup

    def update_header(self, copy_node, transaction, i, freq=1):
        """
        Updates the value of the item in the header table of the tree

        If the item is already present in the header table, it updates the
        linked list of the item,
        else creates a new entry corresponding to the item

        Also updates the support count of the item in the tree
        """
        traveller = None

        if transaction[i] in self.header.keys():
            traveller = self.header[transaction[i]]

            while traveller.next_occurence != None:
                traveller = traveller.next_occurence

            traveller.next_occurence = copy_node
            self.support_count[transaction[i]] += freq
        else:
            self.header[transaction[i]] = copy_node
            self.support_count[transaction[i]] = freq

    def insert_transaction(self, transaction, freq=1):
        """
        Inserts the transaction in the tree.
        `freq` denotes the frequency of the entire transaction

        This function first starts from the root and then tries to find the 
        longest match to the transaction in the tree (updating the support 
        count of the matched items), then adds to the tree if any unmatched 
        items remain in the transaction
        """
        copy_node = self.root

        for i in range(0, len(transaction)):
            found = False
            for next_node in copy_node.next:
                if transaction[i] == next_node.val:
                    found = True
                    next_node.freq += freq
                    copy_node = next_node
                    self.support_count[transaction[i]] += freq
                    break

            if found == False:
                while i < len(transaction):
                    copy_node.next.append(Node(transaction[i], freq))
                    copy_node.next[-1].parent = copy_node
                    copy_node = copy_node.next[-1]
                    self.update_header(copy_node, transaction, i, freq)
                    i += 1

                return True

    def generate_conditional_fp_tree(self, item):
        """
        Generates the conditional FP Tree for the given item
        
        This function is internally called by `mine_fp_tree`

        Returns a Tree which is the consitional FP Tree on the item

        - Uses the header table to find the instances of the item in the tree
        - Then generates mock transactions containing all the nodes from the leaf
        to the root
        - Then creates a new tree
        and adds all these transactions into the new Tree

        - After this, if support count of some items is less than the minimum
        support defined, removes the nodes from the tree
        """

        new_tree = Tree(min_sup=self.min_support)
        copy_node = self.header[item]

        while copy_node != None:
            new_transaction = []
            traveller = copy_node.parent
            new_freq = copy_node.freq
            while traveller.parent != None:
                new_transaction.append(traveller.val)
                traveller = traveller.parent

            if len(new_transaction) > 0:
                new_transaction.reverse()
                new_tree.insert_transaction(new_transaction, new_freq)

            copy_node = copy_node.next_occurence

        items_to_remove = []
        for item in new_tree.support_count.keys():
            if new_tree.support_count[item] < self.min_support:
                items_to_remove.append(item)

        for item in new_tree.header.keys():
            if item in items_to_remove:
                traveller = new_tree.header[item]
                while traveller != None:
                    traveller.parent.next.remove(traveller)
                    traveller.parent.next.extend(traveller.next)
                    traveller = traveller.next_occurence
        
        for item in items_to_remove:
            new_tree.header.pop(item)

        # pprint(self.support_count)
        # print(self.min_support)
        # print("Removed: ", items_to_remove)

        return new_tree

    def mine_fp_tree(self, transaction, freq_itemsets, support_count={}):
        """
        Mines the FP Tree for each frequent item of the tree.

        Returns a dictionary with itemsets as the key and their support
        counts as the values
        
        - For each item in the tree, add it to a temporary transaction list
        - Add this transaction list to the dictionary of frequent itemsets
        - Generate the conditional FP Tree for this item
        - If the tree is not empty, call this function on that tree, passing the 
            original support count as the parameter (reqd to sort) 
        - Else, return
        """
        items = list(self.header.keys())
        items.reverse()
        if support_count == {}:
            support_count = self.support_count
        # print(transaction, end=": ")
        # pprint(items)
        for item in items:
            temp_transaction = []
            temp_transaction.extend(transaction)
            cond_tree = self.generate_conditional_fp_tree(item)

            temp_transaction.append(item)
            temp_transaction = sorted(temp_transaction, key=(lambda x: support_count[x]), reverse=True)
            freq_itemsets[tuple(temp_transaction)] = self.support_count[item]

            # To debug:
            # if temp_transaction == ['d', 'b']:
            #     print("Cond tree of ['d', 'b']")
            #     dft(cond_tree.root)
            #     pprint(cond_tree.header)
            #     pprint(cond_tree.support_count)
            if len(cond_tree.root.next) != 0:
                freq_itemsets = cond_tree.mine_fp_tree(temp_transaction, freq_itemsets, support_count)

        return freq_itemsets


def dft(obj, level=0):
    """
    Depth first traversal of `obj`, which is generally an FP-Tree
    
    Outputs of the form `values:depth:frequency`
    """
    print(str(obj.val)+":"+str(level)+":"+str(obj.freq), end=" ")

    if len(obj.next) == 0:
        print()
    else:
        for i in range(0, len(obj.next)):
            dft(obj.next[i], level+1)


## Generating the Frequent Itemsets

- We now prepare the dataset, keeping only the items which satisfy the minimum support threshold in our dataset
- Then we create an FP Tree using the items in this dataset
- Then we generate frequent itemsets by mining this FP Tree

from pprint import pprint

support_count = {}
min_support = 50

# Determine the support count
for transaction in db:
    for item in transaction:
        if item in support_count.keys():
            support_count[item] += 1
        else:
            support_count[item] = 1

freq_items = {}

# Store the frequent items with their support count in a dictionary
for item in support_count.keys():
    if support_count[item] >= min_support:
        freq_items[item] = support_count[item]

# Sort this dictionary with support count as the key
{k: v for k, v in sorted(freq_items.items(), key=lambda x: x[1], reverse=True)}

# Prepare the dataset in the following way
# for each transaction, intersect it with the set of frequent items
# then sort the remaining items of the transaction based on the support counts
new_db = []
freq_item_list = set(freq_items.keys())

for transaction in db:
    transaction = list(set(transaction).intersection(freq_item_list))
    transaction = sorted(transaction, key=lambda x: freq_items[x], reverse=True)
    new_db.append(transaction)

db = new_db

# Create an FP Tree with the minimum support and the prepared dataset
fp_tree = Tree(min_support)
for transaction in db:
    fp_tree.insert_transaction(transaction)

# Mine the FP tree
frequent_itemsets = fp_tree.mine_fp_tree([], {})
# How many frequent itemsets are generated?
print("No. of frequent itemsets generated:", len(frequent_itemsets))

We now collect all frequent itemsets of length 5 

useful_freq_itemsets = []
for i in frequent_itemsets.keys():
    if len(i) == 5:
        useful_freq_itemsets.append(i)

print("No. of frequent itemsets of length 5:", len(useful_freq_itemsets))

## Rule Generation

We now generate rules based on the frequent itemsets.

In our rules, we have _Yield_ and _Crops_ in the consequent and _Rainfall_, _Soil Type_ and _Season_ as the antecedent

import re

rules = []
for i in useful_freq_itemsets:
    consequent = []
    antecedent = []
    for j in i:
        if len(re.findall(r"._Yield", j)) > 0 or j in crops:
            consequent.append(j)
        else:
            antecedent.append(j)

    rules.append((antecedent, consequent))
print("No. of for rules generated:", len(rules))

We now calculate confidence of each rule

for k in range(0, len(rules)):
    antecedent, consequent = set(rules[k][0]), set(rules[k][1])
    sup_ante, sup_tot = 0, 0
    for i in useful_freq_itemsets:
        if antecedent.issubset(set(i)):
            sup_ante += frequent_itemsets[tuple(i)]
        if antecedent.union(consequent).issubset(set(i)):
            sup_tot += frequent_itemsets[tuple(i)]
            
    confidence = sup_tot/sup_ante
    rules[k] = (rules[k][0], rules[k][1], confidence)
rules[0:5]

We now define a mapping which eases our work of sorting our results on the basis of yield 

mapping = {}
mapping['Very_Low_Yield'] = 0
mapping['Low_Yield'] = 1
mapping['Medium_Yield'] = 2
mapping['High_Yield'] = 3
mapping['Very_High_Yield'] = 4

The `give_crops` function requires the season, soil type and the rainfall as the inputs and it gives a set of suggested crops sorted on the basis of the yield.

We may take only the _very high_, _high_ and _medium_ yields for suggestions based on the number of results returned

def give_crops(season="kharif", soil_type="alluvial", rainfall="medium_rainfall", min_confidence=0):    
    lis = [season, soil_type, rainfall]
    found_set = []
    for rule in rules:
        comp = []
        if rule[2] < min_confidence:
            continue
        for i in rule[0]:
            comp.append(i.lower())
        if set(comp) == set(lis):
            found_set.append([rule[1], rule[2]])
    

    return found_set

Let us now see the recommendations for the crops given by the function

# Test example
# kharif
# alluvial
# medium_rainfall
crop_recommendations = give_crops("rabi", "alluvial", "very_low_rainfall", 0)
pprint(crop_recommendations)

We now discard the _low_yield_ and _very_low_yield_ recommendations and sort the set according to the confidence rules

def give_recommendations(crop_recommendations):
    remove_discr = {}
    
    for r in crop_recommendations:
        if r[0][1] not in remove_discr.keys():
            remove_discr[r[0][1]] = (r[0][0], r[1])
        else:
            if remove_discr[r[0][1]][1] < r[1]:
                remove_discr[r[0][1]] = (r[0][0], r[1])
    
    crop_recommendations = [[[remove_discr[k][0], k], remove_discr[k][1]] for k in remove_discr.keys()]
    
    vh = []
    h = []
    m = []
    for r in crop_recommendations:
        if "Very_High" in r[0][0]:
            vh.append((r[0][1], r[1]))
        elif "High" in r[0][0]:
            h.append((r[0][1], r[1]))
        elif "Medium" in r[0][0]:
            m.append((r[0][1], r[1]))

    vh = sorted(vh, key=lambda x: x[1], reverse=True)
    h = sorted(h, key=lambda x:x[1], reverse=True)
    m = sorted(m, key=lambda x:x[1], reverse=True)
    
    # print(vh, h, m)

    final_suggestion = []
    if len(vh)+len(h)+len(m) == 0:
        print()
        print("Sorry, no crops to suggest.")
        return
    elif len(vh)+len(h) < 5:
        final_suggestion = [i[0] for i in vh]
        final_suggestion.extend([i[0] for i in h])
        final_suggestion.extend(i[0] for i in m)
    else:
        final_suggestion = [i[0] for i in vh]
        final_suggestion.extend([i[0] for i in h])
    
    print()
    print("Suggested Crops:", final_suggestion)

We can now test on more combinations of inputs

crops = give_crops("rabi", "alluvial", "very_low_rainfall", 0.02)
give_recommendations(crops)

crops = give_crops("kharif", "alluvial", "medium_rainfall", 0.02)
give_recommendations(crops)

crops = give_crops("kharif", "red", "very_low_rainfall", 0.02)
give_recommendations(crops)

crops = give_crops("kharif", "arid", "very_low_rainfall", 0.02)
give_recommendations(crops)
